Changed all of the configuration items.
I've updated the configuration directory so that we provide templates
that need to be created before running setup.py install.  These
templates will make setting up a pypi release a bit more complicated and
we'll need to determine if it's worth it or not.

For now, I believe we can start focusing on actual deployments and
seeing if they work correctly.